cambios finales

+++ abmBackend/routes/api.php
@@ -14,22 +14,23 @@ Route::post('/refresh', [AuthController::class, 'refresh'])
     ->middleware('auth:api');
 
 // Rutas protegidas con JWT
-Route::middleware('auth:api')->group(function () {
 
     // Logout
-    Route::post('/logout', [AuthController::class, 'logout']);
+    Route::post('/logout', [AuthController::class, 'logout'])
+        ->middleware('auth:api');
 
     // Endpoints de tareas
-    Route::get('/tasks', [TaskController::class, 'index']);
-    Route::get('/tasks/{id}', [TaskController::class, 'show']);
-    Route::post('/tasks', [TaskController::class, 'store']);
-    Route::put('/tasks/{id}', [TaskController::class, 'update']);
-    Route::delete('/tasks/{id}', [TaskController::class, 'destroy']);
-    Route::post('/tasks/{id}/confirm', [TaskController::class, 'confirmTask']);
+    Route::get('/tasks', [TaskController::class, 'index'])
+        ->middleware('auth:api');
+    Route::get('/tasks/{id}', [TaskController::class, 'show'])
+    ->middleware('auth:api');
+    Route::post('/tasks', [TaskController::class, 'store'])->middleware('auth:api');;
+    Route::put('/tasks/{id}', [TaskController::class, 'update'])->middleware('auth:api');;
+    Route::delete('/tasks/{id}', [TaskController::class, 'destroy'])->middleware('auth:api');;
+    Route::post('/tasks/{id}/confirm', [TaskController::class, 'confirmTask'])->middleware('auth:api');;
 
 
     // Endpoints de usuarios
-    Route::get('/users', [UserController::class, 'index']);
-    Route::get('/users/{id}/tasks', [UserController::class, 'userTasks']);
-    Route::delete('/users/{id}', [UserController::class, 'destroyUser']);
-});
+    Route::get('/users', [UserController::class, 'index'])->middleware('auth:api');;
+    Route::get('/users/{id}/tasks', [UserController::class, 'userTasks'])->middleware('auth:api');;
+    Route::delete('/users/{id}', [UserController::class, 'destroyUser'])->middleware('auth:api');;
